\name{beetle_angle_degrees}
\alias{beetle_angle_degrees}
\title{Convert beetle angle units to degrees}
\usage{beetle_angle_degrees(angle = NULL, species = NULL)}
\arguments{\item{angle}{Optional numeric vector of angle values in dataset units.}\item{species}{Optional character vector of species labels.}}
\value{A numeric vector of angles in degrees.}
\description{Converts angle values from dataset units to degrees, either from a numeric input or from filtered species data.}
