% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{beetle_counts}
\alias{beetle_counts}
\title{Count observations by species}
\usage{
beetle_counts()
}
\value{
A tibble with columns \code{Species} and \code{Count}.
}
\description{
Counts the number of beetles in each species group.
}
