% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{beetle_filter}
\alias{beetle_filter}
\title{Filter beetles by species}
\usage{
beetle_filter(species = NULL)
}
\arguments{
\item{species}{Optional character vector containing one or more of \code{"Con"},
\code{"Hei"}, and \code{"Hep"}.}
}
\value{
A tibble containing the filtered rows.
}
\description{
Filters the beetles dataset to one or more species.
}
