% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{beetle_species}
\alias{beetle_species}
\title{List available species labels}
\usage{
beetle_species()
}
\value{
A character vector of species labels.
}
\description{
Returns the species abbreviations used in the beetles dataset.
}
\examples{
beetle_species()

}
