% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{entomology-package}
\alias{entomology}
\alias{entomology-package}
\title{entomology: simple teaching helpers for flea beetle data}
\description{
Teaching-oriented functions for exploring the flea beetles data used in the
Wisconsin beetle example. The package includes small, readable functions for
filtering, summaries, angle conversion, correlation, and a simple
nearest-centroid classifier.
}
\author{
\strong{Maintainer}: Kelly Bodwin \email{kelly@bodwin.us}

}
\keyword{internal}
